Refactor code to make it adhere to more OOP principles
